Performance tweaks to setIntersects
```
name                           old time/op    new time/op    delta
SetIntersects/with_globs-8       3.05µs ± 1%    0.52µs ±55%   -83.08%  (p=0.008 n=5+5)
SetIntersects/without_globs-8    5.27µs ± 1%    0.26µs ± 1%   -95.05%  (p=0.008 n=5+5)

name                           old alloc/op   new alloc/op   delta
SetIntersects/with_globs-8         276B ± 0%       16B ± 0%   -94.19%  (p=0.008 n=5+5)
SetIntersects/without_globs-8      480B ± 0%        0B       -100.00%  (p=0.008 n=5+5)

name                           old allocs/op  new allocs/op  delta
SetIntersects/with_globs-8         17.0 ± 0%       1.0 ± 0%   -94.12%  (p=0.008 n=5+5)
SetIntersects/without_globs-8      30.0 ± 0%       0.0       -100.00%  (p=0.008 n=5+5)
```
